missing commit

+++ remoteserv/src/main/java/microboot/ComputatorService.java
@@ -1,10 +1,12 @@
 package microboot;
 
+import java.math.BigDecimal;
+
 /**
  * Created by stefanbaychev on 3/24/17.
  */
 public interface ComputatorService {
 
-    Double compute(Double numValueOne, Double numValueTwo);
+    BigDecimal compute(Double numValueOne, Double numValueTwo);
 
 }
